+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix an error when a service use another service
 * Add ERB support in YAML loader
+* Add file_path option
 
 # 0.3.1 (2013-09-17)